Conflicts:
	build.gradle

+++ src/main/java/uristqwerty/CraftGuide/recipes/IC2Power.java
@@ -16,7 +16,7 @@ public class IC2Power implements StackInfoSource
 		}
 		else
 		{
-			int power = Info.itemEnergy.getEnergyValue(itemStack);
+			double power = Info.itemEnergy.getEnergyValue(itemStack);
 
 			if(power > 0)
 			{
